# openapi.yaml
openapi: 3.0.0
info:
    title: RE Remake Interactive Map API
    description: |
        RESTful API for querying Resident Evil Remake game data including items, biohazards, maps, areas, and rooms.

        All endpoints support filtering by various parameters and include rate limiting (100 requests per 15 minutes).

        ## Features
        - Query resources by IDs with batch support
        - Search and filter by multiple criteria
        - S.T.A.R.S. biohazard rankings
        - Comprehensive room and map data

        ## Rate Limiting
        All endpoints are rate-limited to 100 requests per 15-minute window. Rate limit information is provided in response headers.
    version: 1.0.0
    contact:
        name: RE Remake Interactive Map
        url: https://github.com/UncaughtTypeError/RE-Remake-Interactive-Map
servers:
    - url: http://localhost:3000
      description: Local development server
    - url: https://your-production-url.com
      description: Production server
...
paths:
...
    /api/items/all:
        get:
            tags: [Items]
            summary: Get all items (base data)
            description: Retrieves all items base data without any filters. Returns the complete item database with master item definitions.
            responses:
                '200':
                    description: Successful response with all base items
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ItemData'
                '429':
                    $ref: '#/components/responses/RateLimitError'
...
    /api/items/search:
        get:
            tags: [Items]
            summary: Search and filter items (base data)
            description: |
                Search base items by type or name. Returns master item definitions without room locations or difficulty information.
                All parameters are optional and can be combined. Returns items matching ALL specified criteria (AND logic).
            parameters:
                - name: type
                  in: query
                  schema:
                      type: string
                      enum:
                          [
                              Typewriter,
                              ItemBox,
                              Kerosene,
                              Map,
                              PersonOfInterest,
                              DoorKey,
                              InkRibbon,
                              ItemOfInterest,
                              Document,
                              GreenHerb,
                              RedHerb,
                              BlueHerb,
                              FirstAidSpray,
                              SelfDefense,
                              Ammunition,
                              Weapon,
                          ]
                      example: 'Weapon'
                  description: Filter by item type
                - name: name
                  in: query
                  schema:
                      type: string
                      example: 'Handgun'
                  description: Filter by item name (partial match, case-insensitive)
            responses:
                '200':
                    description: Successful response with filtered base items
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ItemData'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '429':
                    $ref: '#/components/responses/RateLimitError'
...
    /api/items/rooms:
        get:
            tags: [Items]
            summary: Get items by IDs (room data)
            description: |
                Retrieves items by their unique IDs with complete room location data. Supports comma-separated batch queries (max 100 IDs).
                Returns both found items and any unrecognized IDs. This returns room-specific instances with location and difficulty information.
            parameters:
                - name: ids
                  in: query
                  required: false
                  schema:
                      type: string
                      example: 'selfDefenseJV-keepersRoom,document-keepersRoom'
                  description: Comma-separated list of item instance IDs (format - itemType-roomId). Max 100 IDs.
                - name: difficulty
                  in: query
                  schema:
                      type: string
                      enum:
                          [
                              JV-lvl-very-easy,
                              JV-lvl-easy,
                              JV-lvl-normal,
                              JV-lvl-hard,
                              CR-lvl-very-easy,
                              CR-lvl-easy,
                              CR-lvl-normal,
                              CR-lvl-hard,
                          ]
                      example: 'JV-lvl-normal'
                  description: Optional - Filter by difficulty level to check availability
            responses:
                '200':
                    description: Successful response with found room items and unrecognized IDs
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    foundItems:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/ItemRoomData'
                                    unrecognizedIds:
                                        type: array
                                        items:
                                            type: string
                                        description: IDs that were not found in the database
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFoundError'
                '429':
                    $ref: '#/components/responses/RateLimitError'
    /api/items/rooms/search:
        get:
            tags: [Items]
            summary: Search and filter items (room data)
            description: |
                Search items with complete room location data by room, difficulty level, type, or name. All parameters are optional and can be combined.
                Returns items matching ALL specified criteria (AND logic). Includes room IDs, map IDs, and difficulty availability.
            parameters:
                - name: room
                  in: query
                  schema:
                      type: string
                      example: 'keepersRoom'
                  description: Filter by room ID (e.g., 'keepersRoom', 'diningRoomF1', 'mainHallF1')
                - name: difficulty
                  in: query
                  schema:
                      type: string
                      enum:
                          [
                              JV-lvl-very-easy,
                              JV-lvl-easy,
                              JV-lvl-normal,
                              JV-lvl-hard,
                              CR-lvl-very-easy,
                              CR-lvl-easy,
                              CR-lvl-normal,
                              CR-lvl-hard,
                          ]
                      example: 'JV-lvl-normal'
                  description: Filter by difficulty level (character and difficulty)
                - name: type
                  in: query
                  schema:
                      type: string
                      enum:
                          [
                              Typewriter,
                              ItemBox,
                              Kerosene,
                              Map,
                              PersonOfInterest,
                              DoorKey,
                              InkRibbon,
                              ItemOfInterest,
                              Document,
                              GreenHerb,
                              RedHerb,
                              BlueHerb,
                              FirstAidSpray,
                              SelfDefense,
                              Ammunition,
                              Weapon,
                          ]
                      example: 'SelfDefense'
                  description: Filter by item type
                - name: name
                  in: query
                  schema:
                      type: string
                      example: 'Battery Pack'
                  description: Filter by item name (partial match, case-insensitive)
            responses:
                '200':
                    description: Successful response with filtered room items
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ItemRoomData'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '429':
                    $ref: '#/components/responses/RateLimitError'

    # ============================================================================
    # PERSONS ENDPOINTS
    # ============================================================================
    /api/persons/all:
        get:
            tags: [Persons]
            summary: Get all persons (base data)
            description: Retrieves all person data without any filters. Returns the complete person database with character information including S.T.A.R.S. team members.
            responses:
                '200':
                    description: Successful response with all persons
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/PersonData'
                            example:
                                - id: 'jillValentine'
                                  name: 'Jill Valentine'
                                  type: 'PersonOfInterest'
                                  taxonomy: 'Persons'
                                  imageSrc: 'https://placeholder-image-url.jpg'
                                  bio: 'to be added'
                                  exclusive: null
                                - id: 'barryBurton'
                                  name: 'Barry Burton'
                                  type: 'PersonOfInterest'
                                  taxonomy: 'Persons'
                                  imageSrc: 'https://placeholder-image-url.jpg'
                                  bio: 'to be added'
                                  exclusive: 'JV'
                                - id: 'albertWesker'
                                  name: 'Albert Wesker'
                                  type: 'PersonOfInterest'
                                  taxonomy: 'Persons'
                                  imageSrc: 'https://placeholder-image-url.jpg'
                                  bio: 'to be added'
                                  exclusive: null
                '429':
                    $ref: '#/components/responses/RateLimitError'
...
components:
    schemas:
        ItemData:
            type: object
            description: Base item data (master item definition)
            properties:
                id:
                    type: string
                    description: Unique item identifier
                    example: 'typewriter'
                name:
                    type: string
                    description: Display name of the item
                    example: 'Typewriter'
                type:
                    type: string
                    description: Item type/category
                    example: 'Typewriter'
                taxonomy:
                    type: string
                    description: Item taxonomy classification
                    example: 'Key Items'
                function:
                    type: string
                    nullable: true
                    description: Item function description (for Typewriter, ItemBox, Kerosene only)
                    example: 'Save Point'
                imageSrc:
                    type: string
                    description: URL to item image
                    example: 'https://s-media-cache-ak0.pinimg.com/originals/f1/26/57/f126578050920fc271fd8b33534a9605.jpg'
                description:
                    type: string
                    description: Item description
                    example: 'A typewriter that can be used to rudimentarily record progression when providing an Ink Ribbon.'
                exclusive:
                    type: string
                    nullable: true
                    description: Character exclusivity (JV/CR) if applicable
                    example: null
        PersonData:
            type: object
            description: Person data for S.T.A.R.S. team members and characters
            properties:
                id:
                    type: string
                    description: Unique person identifier (PersonID)
                    example: 'jillValentine'
                name:
                    type: string
                    description: Display name of the person (PersonDisplayName)
                    example: 'Jill Valentine'
                type:
                    type: string
                    description: Item type (always PersonOfInterest for persons)
                    example: 'PersonOfInterest'
                    enum: [PersonOfInterest]
                taxonomy:
                    type: string
                    description: Taxonomy classification (always Persons)
                    example: 'Persons'
                    enum: [Persons]
                imageSrc:
                    type: string
                    description: URL to person image
                    example: 'https://placeholder-image-url.jpg'
                bio:
                    type: string
                    description: Biography or description of the person
                    example: 'to be added'
                exclusive:
                    type: string
                    nullable: true
                    description: Character exclusivity (JV for Jill Valentine campaign, CR for Chris Redfield campaign, null if appears in both)
                    example: 'JV'
...
        ItemRoomData:
            type: object
            description: Item data with room location information
            properties:
                id:
                    type: string
                    description: Unique item instance identifier (format - itemType-roomId)
                    example: 'selfDefenseJV-keepersRoom'
                itemId:
                    type: string
                    description: Reference to ItemData.id
                    example: 'batteryPack'
                name:
                    type: string
                    description: Display name of the item
                    example: 'Battery Pack'
                type:
                    type: string
                    description: Item type/category
                    example: 'SelfDefense'
                qty:
                    type: integer
                    description: Quantity available
                    example: 1
                qtyItems:
                    type: integer
                    nullable: true
                    description: Quantity group identifier
                    example: null
                uses:
                    type: integer
                    nullable: true
                    description: Number of uses for this item
                    example: null
                isFeatured:
                    type: boolean
                    description: Whether this item is featured
                    example: false
                requirement:
                    type: string
                    description: Requirement to obtain item
                    example: 'none'
                map:
                    type: object
                    properties:
                        roomId:
                            type: string
                            example: 'keepersRoom'
                        mapId:
                            type: string
                            example: 'mansionF1'
                difficultyLevel:
                    type: array
                    items:
                        type: string
                        enum:
                            [
                                JV-lvl-very-easy,
                                JV-lvl-easy,
                                JV-lvl-normal,
                                JV-lvl-hard,
                                CR-lvl-very-easy,
                                CR-lvl-easy,
                                CR-lvl-normal,
                                CR-lvl-hard,
                            ]
                    description: Difficulty levels where this item appears
                    example: ['JV-lvl-very-easy', 'JV-lvl-easy', 'JV-lvl-normal', 'JV-lvl-hard']
...
        Error:
            type: object
            properties:
                error:
                    type: string
                    description: Error message
                    example: 'Invalid query parameters'
                details:
                    type: array
                    items:
                        type: string
                    description: Detailed error information
                    example: ['ids must be comma-separated alphanumeric strings']
    responses:
        ValidationError:
            description: Invalid request parameters
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
                    example:
                        error: 'Invalid query parameters'
                        details: ['ids must be comma-separated alphanumeric strings']
        NotFoundError:
            description: Resource not found
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
                    example:
                        error: 'No IDs provided are recognized'
        RateLimitError:
            description: Rate limit exceeded
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            message:
                                type: string
                                example: 'Too many requests, please try again later.'